Save work in progress fixing term context behavior/bugs
